Fix custom urql provider for express
