+++ lib/graph/use-query.js
@@ -1,9 +1,9 @@
-import { useContext, useState, useEffect } from 'react';
-import { pipe, subscribe } from 'wonka';
-import { createRequest } from 'urql';
-import { UrqlOrderContext } from 'components/urql-order-context';
+const { useContext, useState, useEffect } = require('react');
+const { pipe, subscribe } = require('wonka');
+const { createRequest } = require('urql');
+const { UrqlOrderContext } = require('../../components/urql-order-context');
 
-export const customClientUseQuery = ({ query, variables }) => {
+module.exports = ({ query, variables }) => {
   const [result, setResult] = useState({
     fetching: false,
     error: undefined,
