+++ pages/_app.js
@@ -6,7 +6,7 @@ import { Provider } from 'urql';
 import withUrqlClient from 'lib/with-urql-client';
 import { useMenuAndTenantQuery } from 'lib/graph';
 import AuthGate from 'components/auth-context';
-import UrqlOrderProvider from 'components/urql-order-context';
+import UrqlOrderProvider from 'components/urql-order-provider';
 import BasketProvider from 'components/basket-provider';
 import GlobalStyle from 'ui/global';
